+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [Unreleased] 
+### Added
+- **PODAAC-2783**
+  - Added following Message Attributes to CNMResponse
+     COLLECTION_SHORT_NAME
+     CNM_RESPONSE_STATUS
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
 # [v1.1.1] - 2020-11-19
 ### Added
 ### Changed