+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [v1.2.0] - 2020-12-15
+### Added
+- **PODAAC-2783**
+  - Added following SNS Message Attributes to CNMResponse
+     - COLLECTION
+     - CNM_RESPONSE_STATUS
+     - DATA_VERSION
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+- **Snyk**
+  - Upgrade com.amazonaws:aws-java-sdk-core:1.11.660 -> 1.11.903
+  - Upgrade com.amazonaws:aws-java-sdk-kinesis:1.11.439 -> 1.11.903
+  - Upgrade com.amazonaws:aws-java-sdk-sns:1.11.342 -> 1.11.903
+  
 # [v1.1.1] - 2020-11-19
 ### Added
 ### Changed